# feat(hogql): intelligent argmax on persons table