## Problem

We now have two versions of the persons query argmax grouping. Sometimes one version is faster, sometimes the other.

## Changes

I had a look at one of our bigger clients on EU. I ran a query such as `select id, is_identified, pdi.distinct_id from persons limit 100` and measured the time with both variants, made easy with our [new query debug page](https://eu.posthog.com/debug).

- Old `v1` query. 
  - Selecting `id, properties` - errors / times out
  - Selecting `id, properties.$browser` - errors / times out
  - Selecting `id, is_identified` - 1.6 sec
  - Selecting `id, is_identified, pdi.distinct_id` - 6.5 sec

- New `v2` query. 
  - Selecting `id, properties` - 10.4 sec
  - Selecting `id, properties.$browser` - 10.6 sec
  - Selecting `id, is_identified` - 10.7 sec
  - Selecting `id, is_identified, pdi.distinct_id` - 10.6 sec

Thus, this PR introduces a new default `auto` mode for the `personsArgMaxVersion` modifier, in addition to the existing `v1` and `v2` modes.

If set to `auto`, we will use the `v2` persons argmax query if you are selecting any properties on a person, and the `v1` query if you are not selecting any. This seems to be the fastest case for all.


## How did you test this code?

Wrote a test. Played in the query debugger.